Integrate new publisher vnext work (Preview). * New startup project support * Support SignalR messaging of discovery progress, telemetry, and registry events. * Updated CLI to provide better error descriptions and support signalR event output. * Remove callback support from endpoint and application discovery in favor of signalr * Add new agent framework to write job driven edge agents. * Add new Jobs service and Api * Add new Orchestrator service and Api * Add new preview publisher whose publishing tasks are jobs driven instead of publishednodes.json. * Update publishing API to use new jobs mechanism and factor out into new publisher service. * Support for edge module http request authentication through identity token * Split mime type, content encoding, and event schema into seperate message / Event properties * Rename service project names to convey more what they are doing. * New diagnostics structure to be able to disable app insights if desired. * Only build API nuget packages going forward. * Break apart common projects along the lines of their dependencies, e.g. blob/cosmosdb * Support azuredeploy.json portal and slingshot deployment * Add publisher to opc registry as manageable entity registration * Update scripts * New engineering tool application.

+++ services/deploy/deploy.ps1
@@ -6,7 +6,7 @@
     Deploys the Industrial IoT services dependencies and optionally micro services and UI to Azure.
 
  .PARAMETER type
-    The type of deployment (vm, local)
+    The type of deployment (vm, sim, local)
 
  .PARAMETER applicationName
     The name of the solution. 
@@ -29,8 +29,8 @@
  .PARAMETER containerRegistryPrefix
     Optional, a container registry prefix from which to pull the micro services containers to deploy.
 
- .PARAMETER aadApplicationName
-    The AAD application name to register - defaults to $applicationName. 
+ .PARAMETER applicationRegistrationPrefix
+    The application name prefix for all applications that are registered - defaults to 'iiot' for non-interactive usage. 
 
  .PARAMETER tenantId
     AD tenant to use. 
@@ -38,8 +38,17 @@
  .PARAMETER credentials
     To support non interactive usage of script. (TODO)
 
- .PARAMETER development
-    Whether to deploy for development or production - defaults to $false (production).
+ .PARAMETER serviceAppId
+    The (preregistered) App id of the AAD service application registration.
+
+ .PARAMETER servicePrincipalId
+    The (preregistered) principal Id of the AAD server application
+
+ .PARAMETER clientAppId
+    The (preregistered) App id of the AAD client application registration.
+
+ .PARAMETER clientSecret
+    The (preregistered) App secret for the client.
 
 #>
 
@@ -54,8 +63,11 @@ param(
     [string] $containerRegistryPrefix,
     $credentials,
     [string] $tenantId,
-    [string] $aadApplicationName,
-    [bool] $development = $false,
+    [string] $applicationRegistrationPrefix,
+    [string] $serviceAppId,
+    [string] $servicePrincipalId,
+    [string] $clientAppId,
+    [string] $clientSecret,
     [ValidateSet("AzureCloud")] [string] $environmentName = "AzureCloud"
 )
 
@@ -496,6 +508,20 @@ Function CreateAppRole() {
 # Get configuration object for service and client applications
 #*******************************************************************************************************
 Function GetAzureADApplicationConfig() {
+
+    if (![string]::IsNullOrEmpty($script:serviceAppId) -and `
+        ![string]::IsNullOrEmpty($script:servicePrincipalId)) {
+        Write-Host "Using provided service and client AAD application registrations."
+        return [pscustomobject] @{ 
+            TenantId = $script:tenantId
+            Instance = $script:environment.ActiveDirectoryAuthority
+            ServiceId = $script:serviceAppId
+            ServicePrincipalId = $script:servicePrincipalId
+            ClientId = $script:clientAppId
+            ClientSecret = $script:clientSecret
+        }
+    }
+
     try {
         $creds = ConnectToAzureADTenant
         if (!$creds) {
@@ -506,23 +532,23 @@ Function GetAzureADApplicationConfig() {
             return $null
         }
         
-        $aadApplicationName = $script:aadApplicationName
-        if ([string]::IsNullOrEmpty($aadApplicationName)) {
+        $applicationRegistrationPrefix = $script:applicationRegistrationPrefix
+        if ([string]::IsNullOrEmpty($applicationRegistrationPrefix)) {
             if ($script:interactive) {
                 Write-Host
                 Write-Host "Please provide a name for the AAD application to register."
-                $aadApplicationName = Read-Host "(Hit enter to use '$script:applicationName')"
-                if ([string]::IsNullOrEmpty($aadApplicationName)) {
-                    Write-Host "Registering '$script:applicationName' in AAD."
-                    $aadApplicationName = $script:applicationName
+                $applicationRegistrationPrefix = Read-Host "(Hit enter to use 'iiot')"
+                if ([string]::IsNullOrEmpty($applicationRegistrationPrefix)) {
+                    Write-Host "Registering 'iiot' in AAD."
+                    $applicationRegistrationPrefix = "iiot"
                 }
             }
             else {
-                $aadApplicationName = $script:applicationName
+                $applicationRegistrationPrefix = $script:applicationName
             }
         }
-        $serviceDisplayName = $aadApplicationName + "-services"
-        $clientDisplayName = $aadApplicationName + "-clients"
+        $serviceDisplayName = $applicationRegistrationPrefix + "-service"
+        $clientDisplayName = $applicationRegistrationPrefix + "-client"
 
         $tenant = Get-AzureADTenantDetail
         $tenantName =  ($tenant.VerifiedDomains | Where { $_._Default -eq $True }).Name
@@ -671,20 +697,17 @@ Function GetAzureADApplicationConfig() {
             TenantId = $creds.Tenant.Id
             Instance = $script:environment.ActiveDirectoryAuthority
             Audience = $serviceAadApplication.IdentifierUris[0].ToString()
-            AppName = $aadApplicationName
-            AppId = $serviceAadApplication.AppId
-            AppObjectId = $serviceAadApplication.ObjectId
-            AppDisplayName = $serviceDisplayName
-            ServiceName = $aadApplicationName
             ServiceId = $serviceAadApplication.AppId
-            ServiceObjectId = $serviceAadApplication.ObjectId
+            ClientId = $clientAadApplication.AppId
             ServicePrincipalId = $serviceServicePrincipal.ObjectId
+
+            ServiceObjectId = $serviceAadApplication.ObjectId
             ServiceSecret = $serviceSecret.Value
             ServiceDisplayName = $serviceDisplayName
-            ClientDisplayName = $clientDisplayName
-            ClientId = $clientAadApplication.AppId
             ClientObjectId = $clientAadApplication.ObjectId
             ClientSecret = $clientSecret.Value
+            ClientDisplayName = $clientDisplayName
+
             UserPrincipalId = $user.ObjectId
         }
     }
@@ -719,13 +742,19 @@ Function GetAzureToken() {
         $context = Get-AzureRmContext
         $refreshToken = @($context.TokenCache.ReadItems() | `
             where {$_.tenantId -eq $tenantId -and $_.ExpiresOn -gt (Get-Date)})[0].RefreshToken
+        if (!$refreshToken) {
+            throw "No refresh token found in token cache."
+        }
         $body = "grant_type=refresh_token&refresh_token=$($refreshToken)&resource=74658136-14ec-4630-ad9b-26e160ff0fc6"
-        $apiToken = Invoke-RestMethod "https://login.windows.net/$tenantId/oauth2/token" `
+        $apiToken = Invoke-RestMethod "https://login.windows.net/$($tenantId)/oauth2/token" `
             -Method POST -Body $body -ContentType "application/x-www-form-urlencoded"
+        if (!$apiToken.access_token) {
+            throw "Refreshing token returned empty new token"
+        }
         return $apiToken.access_token
     }
     catch {
-        Write-Host "An error occurred: $($_.Exception.Message)"
+        Write-Host "An error occurred refreshing auth token: $($_.Exception.Message)"
     }
 }
 
@@ -737,19 +766,22 @@ Function GrantPermission() {
         [string] $azureAppId,
         [string] $tenantId
     )
+    $url = "https://main.iam.ad.ext.azure.com/api/RegisteredApplications/$($azureAppId)/Consent?onBehalfOfAll=true"
     try { 
         $token = GetAzureToken -tenantId $tenantId
+        if (!$token) {
+            throw "Failed to get auth token for $($tenantId)."
+        }
         $header = @{
-            "Authorization"          = "Bearer " + $token
+            "Authorization"          = "Bearer $($token)"
             "X-Requested-With"       = "XMLHttpRequest"
             "x-ms-client-request-id" = [guid]::NewGuid()
             "x-ms-correlation-id"    = [guid]::NewGuid()
         } 
-        $url = "https://main.iam.ad.ext.azure.com/api/RegisteredApplications/" + $azureAppId + "/Consent?onBehalfOfAll=true"
         Invoke-RestMethod -Uri $url -Method POST -Headers $header
     }
     catch {
-        Write-Host "An error occurred granting permissions: $($_.Exception.Message)"
+        Write-Host "$($_.Exception.Message) at $($url) requesting consent for your application - Continue..."
     }
 }
 
@@ -864,10 +896,14 @@ catch {
                 Write-Host $_.Exception.Message
             }
             try {
-                Write-Host "Delete AD App "$aadConfig.AppDisplayName
-                Remove-AzureADApplication -ObjectId $aadConfig.AppId
-                Write-Host "Delete AD App "$aadConfig.ClientDisplayName
-                Remove-AzureADApplication -ObjectId $aadConfig.ClientObjectId
+                if (![string]::IsNullOrEmpty($aadConfig.ServiceDisplayName)) {
+                    Write-Host "Delete AD App $($aadConfig.ServiceDisplayName)"
+                    Remove-AzureADApplication -ObjectId $aadConfig.ServiceObjectId
+                }
+                if (![string]::IsNullOrEmpty($aadConfig.ClientDisplayName)) {
+                    Write-Host "Delete AD App $($aadConfig.ClientDisplayName)"
+                    Remove-AzureADApplication -ObjectId $aadConfig.ClientObjectId
+                }
             }
             catch {
                 Write-Host $_.Exception.Message
